pyproject.toml
[project]
name = "pdf-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "fastmcp>=2.2.5",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
